refactor: serve covers as static assets
- Move cover syncing into an explicit script

- Serve copied covers from public assets

- Remove the covers API route and proxy
